fix authtime
