+++ examples/config.json
@@ -9,7 +9,7 @@
     "appid": "simnow_client_test",
     "auth_code": "0000000000000000"
   },
-  "INTERFACE": "ctp_mini",
+  "INTERFACE": "ctp",
   "MD_FUNC": true,
   "TD_FUNC": true,
   "XMIN": [
